+++ convertJs/plc-simulator/src/utils/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Utilities
+ * Exports centralizados
+ */
+
+export * from './fileHandler'